requirements.txt
streamlit
google-generativeai
Pillow
fpdf
requests>=2.28.0
